+++ src/events/FrameRootAddedEvent.hpp
@@ -8,8 +8,8 @@
 #ifndef __FRAME_ROOT_ADDED_EVENT_HPP__
 #define __FRAME_ROOT_ADDED_EVENT_HPP__
 #include "FrameEvent.hpp"
-#include "TransformTreeTypes.hpp"
-#include "Transform.hpp"
+#include <envire_core/TransformTreeTypes.hpp>
+#include <envire_core/Transform.hpp>
 
 namespace envire { namespace core
 {
